+++ CHANGELOG.md
@@ -24,6 +24,8 @@ following conventions:
 
 [KaC]: <https://keepachangelog.com/en/1.0.0/>
 
+## Unreleased
+
 ## 0.2.1.0 (2023-03-21)
 
 ### Breaking