+++ CHANGELOG.md
@@ -24,7 +24,7 @@ following conventions:
 
 [KaC]: <https://keepachangelog.com/en/1.0.0/>
 
-## Unreleased
+## 0.3.0.0 (2023-05-28)
 
 ## Breaking